requirements.txt
bcrypt==5.0.0
cffi==2.0.0
cryptography==46.0.5
graphviz==0.21
invoke==2.2.1
markdown-it-py==4.0.0
mdurl==0.1.2
# Editable install with no version control (ndg==1.0.0)
-e /home/ashraf/code/NDG
netmiko==4.6.0
ntc_templates==9.0.0
paramiko==4.0.0
pycparser==3.0
Pygments==2.19.2
PyNaCl==1.6.2
pyserial==3.5
PyYAML==6.0.3
rich==14.3.3
ruamel.yaml==0.19.1
scp==0.15.0
textfsm==2.1.0
